Enhance TCPMemcachedNodeImpl to know whether auth should happen.
When SASL authentication is requested of a MemcachedClient it
should always ensure that auth takes place before anything
else on the queue is processed.  The best place to handle this
(after experimentation) seems to be at the TCPMemcachedNodeImpl
since the queue and the socket are well known at this level.

It's a bit harder at the BinaryMemcachedNodeImpl, though arguably
makes more sense there.  Since these are all pretty well hidden
as an implementation detail, it's okay for it to be here.

Who knows, maybe someday ascii will have SASL.  Probably not
though.