Recognize operation may be null at times.  e.g.: flush
Both the timeout changes from myself and some of the continuious timeout
changes from Boris assumed there would always be an operation.  In
some cases, like flush, that is not necessarily the case.

Looking at the existing code, there were lots of guards against null
access already, so I just continued that tradition.

The tradition may need to be broken though in the future.

Change-Id: Ic1344ef2df2ab0ba4c03b4e401a4f98436a39772